Final updated version - Everything seems to be fully functional now.
Volkare is added as option in the start menu. He will be added as a City in the CityList.
The "Volkare City" does give -1 rep when attacking it. (hack solution, but it works).
Tooltips are added on the "Start Game" button in the start menu, to explain why it is disabled.

In userActionDialog tooltips are added to the "Enter User Action" button, to explain why it is disabled.

Corrected sorting algorithm for player ranking in GameScoreSystem.cpp, to account for negative scores correctly.

Fixed graph rescaling issues in both positive and negative direction for the y-axis (player score value axis)