New release!
Corrected the hack regarding pointer monster kill sounds array.
The array is now created as a std::vector. This has to be declared as static with immediate size initialization.
This is due to the QSoundEffect library, which does not allow for calls such as .resize() due to a private member in the class library.